Assignment1: create role

+++ Manager/src/main/java/AWSHelper.java
@@ -11,6 +11,7 @@ import software.amazon.awssdk.services.sqs.model.*;
 
 import java.io.File;
 import java.io.InputStream;
+import java.io.UnsupportedEncodingException;
 import java.util.ArrayList;
 import java.util.Base64;
 import java.util.List;
@@ -20,21 +21,21 @@ public class AWSHelper {
     private static Ec2Client ec2 = Ec2Client.create();
     private static SqsClient sqs = SqsClient.builder().region(Defs.REGION).build();
     private static S3Client s3 = S3Client.builder().region(Defs.REGION).build();
-    private static final String amiId = "ami-0742b4e673072066f"; //TODO: CHECK IF GOOD AMI- NEED TO INSTALL SOMETHING ELSE?
+    private static final String amiId = "ami-0a34a1974defc9163";
+    private static final String role = "awsAdminRole";
+    private static final String keyName = "awsKeyPair";
+    private static final String securityGroupIdYOS = "sg-ff035dfe";
+    private static final String securityGroupIdYON = "sg-cb94caca";
 
-
-    //TODO: generailize scripts?
     private static final String managerScript =
-        "#cloud-boothook\n"+
-        "#!/bin/bash\n"+
-        "aws s3 cp s3://assignment1-pre-uploaded-jar/sarcasm.jar sarcasm.jar\n"+
-        "java -jar sarcasm.jar Manager";
+            "#!/bin/bash\n" +
+                    "wget https://assignment1-pre-uploaded-jar.s3.amazonaws.com/Manager.jar\n" +
+                    "java -jar Manager.jar\n";
 
     private static final String workerScript =
-            "#cloud-boothook\n"+
-            "#!/bin/bash\n"+
-            "aws s3 cp s3://assignment1-pre-uploaded-jar/sarcasm.jar sarcasm.jar\n"+
-            "java -jar sarcasm.jar Worker";
+            "#!/bin/bash\n" +
+                    "wget https://assignment1-pre-uploaded-jar.s3.amazonaws.com/Worker.jar\n" +
+                    "java -jar Worker.jar\n";
 
     //EC2
     public static void runManager() {
@@ -46,7 +47,8 @@ public class AWSHelper {
                 System.out.println(id);
                 List<Tag> tags = instance.tags();
                 for (Tag tag : tags) {
-                    if (tag.equals(Defs.MANAGER_TAG) && instance.state().name() != InstanceStateName.TERMINATED) {
+                    if (tag.equals(Defs.MANAGER_TAG) && instance.state().name() != InstanceStateName.TERMINATED
+                            && instance.state().name() != InstanceStateName.SHUTTING_DOWN) {
                         if (instance.state().name() != InstanceStateName.RUNNING &&
                                 instance.state().name() != InstanceStateName.PENDING) {
                             ec2.startInstances(StartInstancesRequest.builder().instanceIds(id).build());
@@ -57,8 +59,12 @@ public class AWSHelper {
             }
         }
         if (!isManager) {
-            createEC2Instance(amiId, managerScript, Defs.MANAGER_TAG);
+            createManagerInstance();
         }
+    }
+
+    public static void createManagerInstance() {
+        createEC2Instance(amiId, managerScript, Defs.MANAGER_TAG);
     }
 
     public static void createWorkerInstance() {
@@ -71,19 +77,32 @@ public class AWSHelper {
         }
     }
 
-    private static void createEC2Instance(String ami, String script, Tag tag) {
+    private static void createEC2Instance(String amiId, String userData, Tag tag) {
+        String base64UserData = null;
+        try {
+            base64UserData = new String(Base64.getEncoder().encode(userData.getBytes("UTF-8")), "UTF-8");
+        } catch (UnsupportedEncodingException e) {
+            e.printStackTrace();
+        }
+
+        IamInstanceProfileSpecification iamSpec = IamInstanceProfileSpecification.builder()
+                .name(role)
+                .build();
+
         RunInstancesRequest runRequest = RunInstancesRequest.builder()
                 .instanceType(InstanceType.T2_MICRO)
                 .imageId(amiId)
-                .maxCount(1)
                 .minCount(1)
-                .userData(Base64.getEncoder().encodeToString(script.getBytes()))
+                .maxCount(1)
+                .keyName(keyName)
+                .securityGroupIds(securityGroupIdYON)
+                .iamInstanceProfile(iamSpec)
+                .userData(base64UserData)
                 .build();
 
         RunInstancesResponse response = ec2.runInstances(runRequest);
         String instanceId = response.instances().get(0).instanceId();
         createTagForInstance(instanceId, tag);
-
     }
 
     private static void createTagForInstance(String id, Tag tag) {
@@ -106,7 +125,8 @@ public class AWSHelper {
                 String id = instance.instanceId();
                 List<Tag> tags = instance.tags();
                 for (Tag tag : tags) {
-                    if (tag.equals(Defs.WORKER_TAG) && instance.state().name() == InstanceStateName.RUNNING) {
+                    if (tag.equals(Defs.WORKER_TAG) && (instance.state().name() == InstanceStateName.RUNNING
+                            || instance.state().name() == InstanceStateName.PENDING)) {
                         activeWorkers++;
                     }
                 }
@@ -122,16 +142,19 @@ public class AWSHelper {
                 String id = instance.instanceId();
                 List<Tag> tags = instance.tags();
                 for (Tag tag : tags) {
-                    if (tag.equals(Defs.WORKER_TAG) && instance.state().name() == InstanceStateName.RUNNING) {
+                    if (tag.equals(Defs.WORKER_TAG) && (instance.state().name() == InstanceStateName.RUNNING
+                    || instance.state().name() == InstanceStateName.PENDING)) {
                         ids.add(id);
                     }
                 }
             }
         }
-        TerminateInstancesRequest request = TerminateInstancesRequest.builder()
-                .instanceIds(ids)
-                .build();
-        ec2.terminateInstances(request);
+        if (!ids.isEmpty()) {
+            TerminateInstancesRequest request = TerminateInstancesRequest.builder()
+                    .instanceIds(ids)
+                    .build();
+            ec2.terminateInstances(request);
+        }
     }
 
     // SQS
@@ -141,6 +164,7 @@ public class AWSHelper {
         initQueue(Defs.WORKER_REQUEST_QUEUE_NAME);
         initQueue(Defs.WORKER_RESPONSE_QUEUE_NAME);
     }
+
     private static void initQueue(String name) {
         try {
             CreateQueueRequest request = CreateQueueRequest.builder()
@@ -150,12 +174,14 @@ public class AWSHelper {
         } catch (QueueNameExistsException e) {
         }
     }
+
     public static String queueUrl(String name) {
         GetQueueUrlRequest getQueueRequest = GetQueueUrlRequest.builder()
                 .queueName(name)
                 .build();
         return sqs.getQueueUrl(getQueueRequest).queueUrl();
     }
+
     public static void sendMessages(String queueName, List<String> bodies) {
         for (String body : bodies) {
             sendMessage(queueName, body);
@@ -166,6 +192,16 @@ public class AWSHelper {
         SendMessageRequest sendMessageRequest = SendMessageRequest.builder()
                 .queueUrl(queueUrl(queueName))
                 .messageBody(body)
+//                .delaySeconds(5)
+                .build();
+        sqs.sendMessage(sendMessageRequest);
+    }
+
+    public static void sendMessageWithDelay(String queueName, String body, int delay) {
+        SendMessageRequest sendMessageRequest = SendMessageRequest.builder()
+                .queueUrl(queueUrl(queueName))
+                .messageBody(body)
+                .delaySeconds(delay)
                 .build();
         sqs.sendMessage(sendMessageRequest);
     }
